requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.105.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
install==1.3.5
joblib==1.4.2
keras==3.4.1
lazy-model==0.2.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
mysql-connector-python==9.0.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pluggy==1.5.0
protobuf==4.25.3
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
rich==13.7.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.24.0.post1
wasabi==1.1.2
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
